test(sessions): keep FSCP03 witness within size limits
